yesdk-567: Anvil-reviewed issues (#348)
* update based on issues discovered by Anvil

* update comments based on review

+++ Yubico.Core/src/Yubico/Core/Devices/Hid/MacOSHidFeatureReportConnection.cs
@@ -68,7 +68,7 @@ namespace Yubico.Core.Devices.Hid
             int deviceEntry = 0;
             try
             {
-                IntPtr matchingDictionary = IORegistryEntryIDMatching(_entryId);
+                IntPtr matchingDictionary = IORegistryEntryIDMatching((ulong)_entryId);
                 deviceEntry = IOServiceGetMatchingService(0, matchingDictionary);
 
                 if (deviceEntry == 0)
